# Removed Django's inefficient icontains from recipient_name