Recipient_name is already upper-case in the matviews

### Example Performance Diff
FY2018, FY2017, FY2016, FY2015, Recipient Lockheed Martin - Table Count
Original: (35.6s, 19.2s, 17.8s)
New: 1.6s